refactor: move view test classes into view package
