+++ src/test/java/core/mvc/view/DummyViewTest.java
@@ -1,4 +1,4 @@
-package core.mvc;
+package core.mvc.view;
 
 import core.mvc.view.DummyView;
 import org.junit.jupiter.api.DisplayName;
